test: Rename test files
