+++ tests/commands/test_check_file_pair.rs
